use std::shared_ptr in the c++ backend

also do some c++ clean up and fix a few c++ test bugs

There are a few more places where we should use shared pointers.

+++ source/lbfgs.cpp
@@ -19,7 +19,7 @@ namespace pele {
         cout << std::setprecision(12);
 
         // allocate space for s_ and y_
-        for (size_t i = 0; i < M_; ++i){
+        for (int i = 0; i < M_; ++i){
             s_.push_back(Array<double>(x_.size()));
             y_.push_back(Array<double>(x_.size()));
         }
